pyproject.toml
[project]
name = "em-database"
version = "0.1.0"
description = "A database for electron microscopy data using pooch to download and manage files."
requires-python = ">=3.12"
dependencies = [
    "pooch",
    "pyyaml",
    "tqdm>=4.67.1",
    "uv>=0.9.27",
]
keywords = [
    "electron microscopy",
    "database",
    "materials science",
]
[project.optional-dependencies]
tests = ["pytest", "pytest-cov", "hyperspy", "quantem"]
doc = ["sphinx", "sphinx-rtd-theme","sphinx-gallery", "quantem", "hyperspy","pydata_sphinx_theme","sphinx_design",]